[Mozilla Branding Removal] Remove branded video from readme
What: Removes the link to an external video demonstrating the use of the Hubs Portals addon from the readme file.

Why: The scene in the video features a large Mozilla logo.

Notes:
It's currently unknown where this video is hosted.

A new version of this video without the Mozilla branding is planned to be added at a later date.